requirements.txt
python-telegram-bot[job-queue]==21.11.1
python-dotenv==1.0.1
pandas==2.2.3
yfinance==0.2.40
pandas-ta==0.4.71b0
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
google-generativeai==0.8.4
